Moved monitor selection list to architecture models tab, added buttons to (de-)select all monitors
