+++ bundles/org.palladiosimulator.simexp.ui.workflow/src/org/palladiosimulator/simexp/ui/workflow/config/SimExpSimulationParameterTab.java
@@ -10,13 +10,8 @@ import org.eclipse.swt.events.ModifyListener;
 import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
-import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.Event;
-import org.eclipse.swt.widgets.Group;
 import org.eclipse.swt.widgets.Label;
-import org.eclipse.swt.widgets.List;
-import org.eclipse.swt.widgets.Listener;
 import org.eclipse.swt.widgets.Text;
 import org.palladiosimulator.simexp.commons.constants.model.SimulationConstants;
 
@@ -30,8 +25,6 @@ public class SimExpSimulationParameterTab extends AbstractLaunchConfigurationTab
 	private Text textSimulationID;
 	private Text textNumberOfRuns;
 	private Text textNumerOfSimulationsPerRun;
-	private List availableMonitors;
-	private List selectedMonitors;
 	
 	private Composite container;
 	private ModifyListener modifyListener;
@@ -68,68 +61,7 @@ public class SimExpSimulationParameterTab extends AbstractLaunchConfigurationTab
         textNumerOfSimulationsPerRun = new Text(container, SWT.BORDER);
         textNumerOfSimulationsPerRun.setLayoutData(new GridData(SWT.FILL, SWT.CENTER, true, false));
         textNumerOfSimulationsPerRun.addModifyListener(modifyListener);
-        
-        createMiddleContent();
 	}
-	
-	private void createMiddleContent() {
-        Composite monitorComposite = new Composite(container, SWT.NONE);
-        monitorComposite.setLayout(new GridLayout(3, false));
-        monitorComposite.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true));
-
-        Group leftGroup = new Group(monitorComposite, SWT.NONE);
-        leftGroup.setText("Available monitors");
-        leftGroup.setLayout(new GridLayout(1, false));
-        leftGroup.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true));
-
-        availableMonitors = new List(leftGroup, SWT.MULTI | SWT.BORDER);
-        availableMonitors.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true));
-
-        Composite buttons = new Composite(monitorComposite, SWT.NONE);
-        buttons.setLayout(new GridLayout(1, false));
-
-        Button moveRight = new Button(buttons, SWT.PUSH);
-        moveRight.setText(">");
-        moveRight.setLayoutData(new GridData(SWT.FILL, SWT.CENTER, true, true));
-        moveRight.addListener(SWT.Selection, new Listener() {
-			@Override
-			public void handleEvent(Event event) {
-				if (event.type == SWT.Selection) {
-					String[] selection = availableMonitors.getSelection();
-					for (String s : selection) {
-						availableMonitors.remove(s);
-						selectedMonitors.add(s);
-					}
-				}
-			}
-		});
-        
-        Button moveLeft = new Button(buttons, SWT.PUSH);
-        moveLeft.setText("<");
-        moveLeft.setLayoutData(new GridData(SWT.FILL, SWT.CENTER, true, true));
-        moveLeft.addListener(SWT.Selection, new Listener() {
-			@Override
-			public void handleEvent(Event event) {
-				if (event.type == SWT.Selection) {
-					String[] selection = selectedMonitors.getSelection();
-					for (String s : selection) {
-						availableMonitors.add(s);
-						selectedMonitors.remove(s);
-					}
-				}
-			}
-		});
-
-        Group rightGroup = new Group(monitorComposite, SWT.NONE);
-        rightGroup.setText("Selected monitors");
-        rightGroup.setLayout(new GridLayout(1, false));
-        rightGroup.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true));
-
-        selectedMonitors = new List(rightGroup, SWT.MULTI | SWT.BORDER);
-        selectedMonitors.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true));
-    }
-	
-	
 
 	@Override
 	public void setDefaults(ILaunchConfigurationWorkingCopy configuration) {
@@ -158,6 +90,8 @@ public class SimExpSimulationParameterTab extends AbstractLaunchConfigurationTab
         } catch (CoreException e) {
             LaunchConfigPlugin.errorLogger(getName(),"Number of simulations per run", e.getMessage());
         }
+		
+		
 	}
 
 	@Override
@@ -177,6 +111,8 @@ public class SimExpSimulationParameterTab extends AbstractLaunchConfigurationTab
 		} catch (NumberFormatException ignore) {
 			
 		}
+		
+		
 		
 	}
